+++ uiza/api_resources/__init__.py
@@ -1,0 +1,3 @@
+from uiza.api_resources.user import *
+from uiza.api_resources.entity import *
+from uiza.api_resources.base import *
