+++ uiza/api_resources/base/__init__.py
@@ -1,0 +1,2 @@
+from uiza.api_resources.base.connections import Connection
+from uiza.exceptions import *
