+++ uiza/api_resources/base/base.py
