+++ uiza/api_resources/entity/__init__.py
@@ -1,0 +1,1 @@
+from uiza.api_resources.entity.entity import Entity
