--- uiza/base/__init__.py
@@ -1,2 +1,0 @@
-from uiza.base.connections import Connection
-from uiza.exceptions import *
