--- uiza/entity/__init__.py
@@ -1,1 +1,0 @@
-from uiza.entity.entity import Entity
