--- uiza/user/__init__.py
@@ -1,1 +1,0 @@
-from uiza.user.user import User
